shared-models module
