+++ event-details/arrival-tracking/ArrivalOperation.ts
@@ -1,0 +1,11 @@
+import {
+  EventArrival,
+  EventArrivalOperationResult
+} from "@shared-models/EventArrivals"
+
+export type EventArrivalOperationKind = "arrived" | "departed"
+
+export type PerformArrivalsOperation = (
+  arrivals: EventArrival[],
+  operation: EventArrivalOperationKind
+) => Promise<EventArrivalOperationResult[]>
