+++ event-details/arrival-tracking/MockData.ts
@@ -1,6 +1,6 @@
 import { faker } from "@faker-js/faker"
 import { mockLocationCoordinate2D } from "@location/MockData"
-import { EventArrival } from "@event-details/arrival-tracking/Models"
+import { EventArrival } from "@shared-models/EventArrivals"
 import { randomFloatInRange } from "@lib/utils/Random"
 
 export const mockEventArrival = (): EventArrival => ({
