+++ location/MockData.ts
@@ -1,5 +1,5 @@
 import { LocationObject, LocationObjectCoords } from "expo-location"
-import { LocationCoordinate2D, TiFLocation } from "./Location"
+import { LocationCoordinate2D, TiFLocation } from "@shared-models/Location"
 import { randomFloatInRange, randomlyNull } from "@lib/utils/Random"
 import { faker } from "@faker-js/faker"
