+++ location/index.ts
@@ -1,5 +1,5 @@
-export * from "./Location"
-export * from "./Placemark"
+export * from "@shared-models/Location"
+export * from "@shared-models/Placemark"
 export * from "./Region"
 export * from "./Geocoding"
 export * from "./UserLocation"
